Fix GetKnownPath with path mappings on Linux
GetKnownPath can return default, if the KnownPath isn't supported on the
platform. Since multiple KnownPath values can produce this result on
Linux, doing a dictionary lookup with default as the key, will return
the same value for those KnownPath values.

This commit adds a new type of mapping: KnownPath mappings, which maps a
KnownPath key to an AbsolutePath value.