Downloader rework (#481)
* Document and extract the structs from the downloader code

* Massive refactoring of the code, pulled helper code out into the classes they're connected with. Performs a single prefetch command for getting the file size, instead of trying to work around some chunks being the "starter" chunks.

Removes a lot of methods that are used in single places and inlines them, reducing the method name bloat.

Now allocates chunks based on chunk sizes, instead of total number of chunks per download.

* I think it finally works, needs more testing though

* Add handling for non-resumable downloads

This update includes a new method that provides a fallback mechanism for when an HTTP server does not support multipart downloads. Now, instead of throwing an exception, the application will attempt to download without resume. In addition, the local HTTP server test has been improved to better handle different HTTP methods and ensure the correct content length.

* WIP

* Fix resume tests

* Test cleanup

* "Refactored size-related variables to Size type"

Replaced every instance of 'long' type for size and offset related variables to 'Size' type in the HttpDownloader. The 'Size' type introduces a more formalized way to represent sizes across the application, thereby enhancing code readability and maintainability. The 'Size' type also provides various helper methods and security checks that weren't available with the 'long' type. This allows for a more standardized and safer way to manipulate the respective variables across different classes and methods.

* More cleanup

* Final cleanup

* Update directory paths in clean_environment_tests workflow

The paths to the 'src' and 'tests' directories in the clean_environment_tests workflow have been updated. The previous paths which started with "./" were incorrect and were preventing the workflow to trigger properly. Now they are corrected to "src/**" and "tests/**" respectively.

* This commit includes several small changes with the primary purpose of enhancing code readability and consistency. Changes include: removing `public` access modifier from DownloadState class as not necessary, renaming several private fields in the AdvancedHttpDownloader class to follow underscore prefix convention for private fields, and reducing the parameters in several methods where possible to enhance clarity and readability. A comment was added in the LocalHttpServer.cs for better understanding of the function. The lack of solid changes in functionality means these refactors should not affect performance or expected results.

* Allow NexusMods.Networking.HttpDownloader to be visible to its Test project.

Preparing the HttpDownloader project for unit testing by adding InternalsVisibleTo attribute in the project file. Implementing LocalHttpServer for serving resource files during tests to simulate real-life scenarios.."

* Add LocalHttpServer to CLI tests