Small Rework: Mod Page Update info now provides info for mods at an individual level. [Read Note]
What this means is that the consumer, for example the update button on the mod page row can now tell apart multiple mods on a single mod page.

This is important, in the case there are updates for multiple mods on a single mod page.

This enables us, as premium to download the update for each mod.

In addition, we will now also be able to determine if we may already have a newer version of an existing file in out datastore with this rework. More code for that in an upcoming commit.