Prevent long-running task from keeping app alive
Related to #2765. Sometimes the app doesn't close properly after closing
the main window. Using the debugger I could see that the main thread was
waiting for the host to be disposed. A long-running task fetching the
user avatar was also blocked on something.

I couldn't get more information out of my debugging session, but my
guess is that something is blocking the host from being disposed
properly.