Update index code to be resilient against duplicate keys
Fixes #2869.

This PR updates `ReindexState` in the Synchronizer to be more resilient
against duplicate keys. I've added checks to make sure we don't index
files we've already seen before which can lead to duplicate disk entry
entities. I've also replaced the usage of `ToDictionary` which throws an
exception on duplicate keys and we're now logging if we received
duplicate entities.

I'm still not 100% how duplicate paths ended up in the disk state, my
guess is that we previously had worse code that added duplicate entries
and we fixed the places where we added duplicates but we never removed
them, so users that upgraded to newer version still had duplicates in
the DB.