Parse ZIP extra data field `0x7075` for proper unicode names
Fixes #3419.

`System.IO.Compression` only supports the general purpose bit 11 for
unicode file names and it doesn't expose a way for us to parse the extra
data fields.

I've decided to swap to `SharpZipLib` instead because it also handles
unicode file names and exposes the extra data fields that we can parse
for the `0x7075` field containing the unicode path for entries that
don't have the general purpose bit 11 set.